Restructure Cards in seperate directory, Redesign
