+++ src/Components/Cards/index.js
@@ -1,0 +1,1 @@
+export { ProductCard } from "./ProductCard";
