+++ src/styles.css
@@ -1,0 +1,8 @@
+@import url('https://fonts.googleapis.com/css2?family=Raleway&display=swap');
+
+html{
+  font-family: 'Raleway', sans-serif;
+}
+.App {
+  text-align: center;
+}
